% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.staple.food.basket-package.R
\docType{package}
\name{dc.staple.food.basket-package}
\alias{dc.staple.food.basket}
\alias{dc.staple.food.basket-package}
\title{dc.staple.food.basket: Social Data Commons: Health, Behavior, Diet: Staple Food Basket}
\description{
Allows user to easily get Staple Food Basket data from the SDAD Data Commons. Data can be provided directly or in a file.
}
\author{
\strong{Maintainer}: Social and Decision Analytics Division (SDAD), Biocomplexity Institute, University of Virginia

}
\keyword{internal}
