% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_file_names.R
\name{get_data_file_names}
\alias{get_data_file_names}
\title{Get Data File Names (from Dataverse)
Retrieve data file names in the Dataverse dataset.}
\usage{
get_data_file_names()
}
\description{
Get Data File Names (from Dataverse)
Retrieve data file names in the Dataverse dataset.
}
\examples{
\dontrun{get_data_file_names()}
}
